add 2 openalpr activities
