+++ FaceTracker/app/CMakeLists.txt
@@ -6,9 +6,9 @@ cmake_minimum_required(VERSION 3.4.1)
 set(CMAKE_VERBOSE_MAKEFILE on)
 set (CMAKE_CXX_STANDARD 11)
 
-include_directories(/home/irina/openalpr_android_build/OpenCV-android-sdk/sdk/native/jni/include)
+include_directories($ENV{OPENCV_ANDROID_SDK}/sdk/native/jni/include)
 add_library( lib_opencv SHARED IMPORTED )
-set_target_properties(lib_opencv PROPERTIES IMPORTED_LOCATION /home/irina/openalpr_android_build/OpenCV-android-sdk/sdk/native/libs/${ANDROID_ABI}/libopencv_java3.so)
+set_target_properties(lib_opencv PROPERTIES IMPORTED_LOCATION $ENV{OPENCV_ANDROID_SDK}/sdk/native/libs/${ANDROID_ABI}/libopencv_java3.so)
 
 find_library( # Sets the name of the path variable.
               log-lib
