+++ FaceTracker/app/build.gradle
@@ -25,8 +25,8 @@ android {
             cmake {
                 cppFlags "-frtti -fexceptions -std=c++11"
                 //abiFilters 'x86', 'x86_64', 'armeabi-v7a', 'arm64-v8a'
-                abiFilters "${cpuType}"
-                //abiFilters 'x86_64', 'armeabi-v7a', 'arm64-v8a'
+                //abiFilters "${cpuType}"
+                abiFilters 'armeabi-v7a', 'arm64-v8a'
                 arguments "-DANDROID_STL=gnustl_shared"
                 //arguments "-DANDROID_STL=c++_shared"
                 // AL:  arguments "-DANDROID_STL=c++_shared", "-DBUILD_opencv_apps:BOOL=ON", "-DBUILD_SHARED_LIBS:BOOL=ON"
