+++ FaceTracker/app/src/main/AndroidManifest.xml
@@ -33,9 +33,9 @@
             android:value="face" />
 
         <activity
-            android:name="tensorflow.detector.spc.CameraActivityALPRMain"
+            android:name="tensorflow.detector.spc.CameraActivityMainALPReu"
             android:icon="@drawable/icon"
-            android:label="@string/app_alpr_name"
+            android:label="@string/app_alpr_name_eu"
             android:theme="@style/AppTheme"
             android:screenOrientation="portrait">
             <intent-filter>
@@ -46,7 +46,20 @@
         </activity>
 
         <activity
-            android:name="tensorflow.detector.spc.CameraActivityMain"
+            android:name="tensorflow.detector.spc.CameraActivityMainALPRus"
+            android:icon="@drawable/icon"
+            android:label="@string/app_alpr_name_us"
+            android:theme="@style/AppTheme"
+            android:screenOrientation="portrait">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+
+        <activity
+            android:name="tensorflow.detector.spc.CameraActivityMainSPC"
             android:icon="@drawable/icon"
             android:label="@string/app_name"
             android:theme="@style/AppTheme"
